move helper to helper file
