+++ ui-components/index.ts
@@ -70,7 +70,7 @@ export * from "./src/helpers/formatIncome"
 export * from "./src/helpers/validators"
 export * from "./src/helpers/blankApplication"
 export * from "./src/helpers/formatYesNoLabel"
-export * from "./src/helpers/getTranslationWithArguments"
+export * from "./src/helpers/translationHelpers"
 export * from "./src/helpers/preferences"
 export * from "./src/helpers/resolveObject"
 export * from "./src/helpers/useMutate"
