+++ ui-components/src/tables/StandardTable.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react"
 import { DragDropContext, Droppable, Draggable, DropResult } from "react-beautiful-dnd"
 import { nanoid } from "nanoid"
-import { getTranslationWithArguments } from "../helpers/getTranslationWithArguments"
+import { getTranslationWithArguments } from "../helpers/translationHelpers"
 import { Icon } from "../icons/Icon"
 import { t } from "../helpers/translator"
